LayoutTests/http/tests/security/dataURL/xss-DENIED-to-data-url-window-open-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

Opener Frame

Test
